+++ CHANGELOG.md
@@ -3,8 +3,7 @@
 	Placeholder for the next version (at the beginning of the line):
 	## __WORK IN PROGRESS__
 -->
-
-## __WORK IN PROGRESS__
+## 0.6.0 (2022-10-02)
 - (Apollon77) Add option --doNotWatch for "watch" mode to ignore changes on defined files or locations
 - (Apollon77/kleinOr) Set several other system settings on "setup" call like the location, currency, language and temperature unit that they are not empty
 - (Apollon77) Only send SIGINT on exit first and then SIGKILL after 5 seconds if not yet exited to allow clean shutdown of all components
@@ -19,7 +18,6 @@
 - (UncleSamSwiss) `dev-server watch` and `debug` support `--noInstall` which won't build/install the adapter
 - (UncleSamSwiss) js-controller is now always started with inspector on port 9228 (#150)
 - (UncleSamSwiss) Improvements during `dev-server setup` (#114, #201)
-
 
 ## 0.4.1 (2022-04-29)